refactor(query): #240 split sort-key resolution into entity_query_sort.go (500-line limit)
